+++ app/src/main/java/com/wearerommies/roomie/presentation/navigator/MainNavigator.kt
@@ -26,7 +26,7 @@ import com.wearerommies.roomie.presentation.ui.mypage.contact.navigation.navigat
 import com.wearerommies.roomie.presentation.ui.mypage.gender.navigation.navigateToGender
 import com.wearerommies.roomie.presentation.ui.mypage.myaccount.navigation.navigateToMyAccount
 import com.wearerommies.roomie.presentation.ui.mypage.name.navigation.navigateToName
-import com.wearerommies.roomie.presentation.ui.mypage.navigation.navigateToMy
+import com.wearerommies.roomie.presentation.ui.mypage.my.navigation.navigateToMy
 import com.wearerommies.roomie.presentation.ui.mypage.nickname.navigation.navigateToNickname
 import com.wearerommies.roomie.presentation.ui.search.navigation.navigateToSearch
 import com.wearerommies.roomie.presentation.ui.tour.navigation.navigateToTourFirstStep
