+++ app/src/main/java/com/wearerommies/roomie/presentation/navigator/component/RoomieNavHost.kt
@@ -18,7 +18,7 @@ import com.wearerommies.roomie.presentation.ui.mypage.contact.navigation.contact
 import com.wearerommies.roomie.presentation.ui.mypage.gender.navigation.genderNavGraph
 import com.wearerommies.roomie.presentation.ui.mypage.myaccount.navigation.myAccountNavGraph
 import com.wearerommies.roomie.presentation.ui.mypage.name.navigation.nameNavGraph
-import com.wearerommies.roomie.presentation.ui.mypage.navigation.myNavGraph
+import com.wearerommies.roomie.presentation.ui.mypage.my.navigation.myNavGraph
 import com.wearerommies.roomie.presentation.ui.mypage.nickname.navigation.nicknameNavGraph
 import com.wearerommies.roomie.presentation.ui.search.navigation.searchNavGraph
 import com.wearerommies.roomie.presentation.ui.splash.navigation.splashNavGraph
