+++ app/src/main/java/com/wearerommies/roomie/presentation/ui/mypage/my/MySideEffect.kt
@@ -1,4 +1,4 @@
-package com.wearerommies.roomie.presentation.ui.mypage
+package com.wearerommies.roomie.presentation.ui.mypage.my
 
 sealed class MySideEffect {
     data class ShowToast(val message: String) : MySideEffect()
