+++ app/src/main/java/com/wearerommies/roomie/presentation/ui/mypage/my/navigation/MyNavigation.kt
@@ -1,4 +1,4 @@
-package com.wearerommies.roomie.presentation.ui.mypage.navigation
+package com.wearerommies.roomie.presentation.ui.mypage.my.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.wearerommies.roomie.presentation.navigator.route.MainTabRoute
-import com.wearerommies.roomie.presentation.ui.mypage.MyRoute
+import com.wearerommies.roomie.presentation.ui.mypage.my.MyRoute
 
 fun NavController.navigateToMy(navOptions: NavOptions) {
     navigate(
